feat(sessions): move init & finalize w/ `MPI_Session` to its own file
